S60: DF64 FMA + transcendentals + polyfill hardening + deep debt + doc cleanup
DF64 core (df64_core.wgsl):
- FMA-optimized two_prod: 17 ops → 2 via fma(a,b,-p), eliminates Dekker split()
- df64_mul cross-terms use FMA for better stability in Krylov solvers

DF64 transcendentals (new df64_transcendentals.wgsl):
- sqrt_df64 (Newton-Raphson), exp_df64 (Cody-Waite + Horner), log_df64 (atanh)
- sin_df64/cos_df64 (Cody-Waite π/2 reduction), pow_df64, tanh_df64
- All at FP32 core speed, no vendor math library dependency

4 force shaders evolved hybrid → full FP32 core streaming:
- born_mayer_df64, morse_df64, yukawa_df64, lennard_jones_df64
- Zero f64-unit transcendental calls; sqrt/exp stay in DF64

Polyfill patcher hardened:
- patch_transcendentals_in_code() protects ldexp/exp_df64/log_df64 from
  substring collision mangling via sentinel-based replacement

Deep debt fixes:
- Crank-Nicolson: variable shadowing bug (Courant r shadowed by Dirichlet(r))
- Cholesky: SPD validation (NaN/non-positive diagonal detection)
- Cross-attention: q_seq_len/kv_seq_len evolution (6 Rust + 6 WGSL files)
- Loop unroller: test assertions corrected for u32 literal suffix
- Multi-GPU: deterministic adapter selection via BARRACUDA_GPU_ADAPTER

Doc cleanup:
- Root docs updated to Session 60 (README, STATUS, NEXT_STEPS, DEBT, DOCUMENTATION)
- Removed 6 stale docs (REPO_STATUS, WGPU_REFACTORING, docs/team/*)
- Removed 4 redundant root docs (ABSORPTION_TRACKER, DEEP_DEBT_STATUS, etc.)

+++ crates/barracuda/src/shaders/math/hill_f64.wgsl
@@ -33,8 +33,8 @@ fn main(@builtin(global_invocation_id) global_id: vec3<u32>) {
     let idx = global_id.x;
     if (idx >= params.n_elements) { return; }
 
-    let x  = max(input[idx], 0.0);
-    let Kn = pow(max(params.K, 1e-30), params.n);
+    let x  = max(input[idx], f64(0.0));
+    let Kn = pow(max(params.K, f64(1e-30)), params.n);
     let xn = pow(x, params.n);
     output[idx] = xn / (Kn + xn);
 }
